added score and graveyard for both sides, changed labels
